+++ crates/ecstore/src/set_disk/core/mod.rs
@@ -1,0 +1,20 @@
+// Copyright 2024 RustFS Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Shared low-level IO/erasure primitives for the `SetDisks` core, extracted
+//! during the God-Object split (P5, tracking backlog#815, issue backlog#820).
+//! Operation families reach these through the `SetDisks` core rather than
+//! duplicating read/write/erasure logic.
+
+pub(crate) mod io_primitives;
